+++ api-client-java/src/main/java/io/daytona/api/client/model/SandboxListItem.java
@@ -140,12 +140,12 @@ public class SandboxListItem {
   public static final String SERIALIZED_NAME_CPU = "cpu";
   @SerializedName(SERIALIZED_NAME_CPU)
   @javax.annotation.Nonnull
-  private BigDecimal cpu;
+  private Integer cpu;
 
   public static final String SERIALIZED_NAME_GPU = "gpu";
   @SerializedName(SERIALIZED_NAME_GPU)
   @javax.annotation.Nonnull
-  private BigDecimal gpu;
+  private Integer gpu;
 
   public static final String SERIALIZED_NAME_SPOT = "spot";
   @SerializedName(SERIALIZED_NAME_SPOT)
@@ -165,12 +165,12 @@ public class SandboxListItem {
   public static final String SERIALIZED_NAME_MEMORY = "memory";
   @SerializedName(SERIALIZED_NAME_MEMORY)
   @javax.annotation.Nonnull
-  private BigDecimal memory;
+  private Integer memory;
 
   public static final String SERIALIZED_NAME_DISK = "disk";
   @SerializedName(SERIALIZED_NAME_DISK)
   @javax.annotation.Nonnull
-  private BigDecimal disk;
+  private Integer disk;
 
   public static final String SERIALIZED_NAME_LABELS = "labels";
   @SerializedName(SERIALIZED_NAME_LABELS)
@@ -604,7 +604,7 @@ public class SandboxListItem {
   }
 
 
-  public SandboxListItem cpu(@javax.annotation.Nonnull BigDecimal cpu) {
+  public SandboxListItem cpu(@javax.annotation.Nonnull Integer cpu) {
     this.cpu = cpu;
     return this;
   }
@@ -614,16 +614,16 @@ public class SandboxListItem {
    * @return cpu
    */
   @javax.annotation.Nonnull
-  public BigDecimal getCpu() {
+  public Integer getCpu() {
     return cpu;
   }
 
-  public void setCpu(@javax.annotation.Nonnull BigDecimal cpu) {
+  public void setCpu(@javax.annotation.Nonnull Integer cpu) {
     this.cpu = cpu;
   }
 
 
-  public SandboxListItem gpu(@javax.annotation.Nonnull BigDecimal gpu) {
+  public SandboxListItem gpu(@javax.annotation.Nonnull Integer gpu) {
     this.gpu = gpu;
     return this;
   }
@@ -633,11 +633,11 @@ public class SandboxListItem {
    * @return gpu
    */
   @javax.annotation.Nonnull
-  public BigDecimal getGpu() {
+  public Integer getGpu() {
     return gpu;
   }
 
-  public void setGpu(@javax.annotation.Nonnull BigDecimal gpu) {
+  public void setGpu(@javax.annotation.Nonnull Integer gpu) {
     this.gpu = gpu;
   }
 
@@ -699,7 +699,7 @@ public class SandboxListItem {
   }
 
 
-  public SandboxListItem memory(@javax.annotation.Nonnull BigDecimal memory) {
+  public SandboxListItem memory(@javax.annotation.Nonnull Integer memory) {
     this.memory = memory;
     return this;
   }
@@ -709,16 +709,16 @@ public class SandboxListItem {
    * @return memory
    */
   @javax.annotation.Nonnull
-  public BigDecimal getMemory() {
+  public Integer getMemory() {
     return memory;
   }
 
-  public void setMemory(@javax.annotation.Nonnull BigDecimal memory) {
+  public void setMemory(@javax.annotation.Nonnull Integer memory) {
     this.memory = memory;
   }
 
 
-  public SandboxListItem disk(@javax.annotation.Nonnull BigDecimal disk) {
+  public SandboxListItem disk(@javax.annotation.Nonnull Integer disk) {
     this.disk = disk;
     return this;
   }
@@ -728,11 +728,11 @@ public class SandboxListItem {
    * @return disk
    */
   @javax.annotation.Nonnull
-  public BigDecimal getDisk() {
+  public Integer getDisk() {
     return disk;
   }
 
-  public void setDisk(@javax.annotation.Nonnull BigDecimal disk) {
+  public void setDisk(@javax.annotation.Nonnull Integer disk) {
     this.disk = disk;
   }
